+++ scripts/build-deps.ps1
@@ -269,7 +269,7 @@ function buildSentryNative()
     Push-Location -Path $NativePath
     try
     {
-        cmake -B "build" -D SENTRY_BACKEND=crashpad -D SENTRY_SDK_NAME=sentry.native.unreal -D SENTRY_BUILD_SHARED_LIBS=OFF -D SENTRY_BATCHER_BUFFER_COUNT=10
+        cmake -B "build" -D SENTRY_BACKEND=crashpad -D SENTRY_SDK_NAME=sentry.native.unreal -D SENTRY_BUILD_SHARED_LIBS=OFF -D SENTRY_BATCHER_BUFFER_COUNT=10 -D SENTRY_HANDLER_STACK_SIZE=128
         cmake --build "build" --target sentry --config RelWithDebInfo --parallel
         cmake --build "build" --target crashpad_handler --config RelWithDebInfo --parallel
         cmake --install "build" --prefix "install" --config RelWithDebInfo
@@ -295,7 +295,7 @@ function buildSentryNative()
     Push-Location -Path $NativePath
     try
     {
-        cmake -B "build_native" -D SENTRY_BACKEND=native -D SENTRY_SDK_NAME=sentry.native.unreal -D SENTRY_BUILD_SHARED_LIBS=OFF -D SENTRY_BATCHER_BUFFER_COUNT=10
+        cmake -B "build_native" -D SENTRY_BACKEND=native -D SENTRY_SDK_NAME=sentry.native.unreal -D SENTRY_BUILD_SHARED_LIBS=OFF -D SENTRY_BATCHER_BUFFER_COUNT=10 -D SENTRY_HANDLER_STACK_SIZE=128
         cmake --build "build_native" --target sentry --config RelWithDebInfo --parallel
         cmake --build "build_native" --target sentry-crash --config RelWithDebInfo --parallel
         cmake --install "build_native" --prefix "install_native" --config RelWithDebInfo