+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## Unreleased
 
+### Features
+
+- Add app-hang tracking via NDK game-thread heartbeat on Android ([#1497](https://github.com/getsentry/sentry-unreal/pull/1497))
+
 ### Fixes
 
+- Increase the crash handler thread stack guarantee to 128 KiB on Windows to make stack-overflow crash capture more reliable ([#1500](https://github.com/getsentry/sentry-unreal/pull/1500))
 - Remove legacy session replay file attachment in favor of Session Replay envelope ([#1493](https://github.com/getsentry/sentry-unreal/pull/1493))
 
 ### Dependencies
@@ -20,6 +25,9 @@
 - Bump Android Gradle Plugin from v6.15.0 to v6.16.0 ([#1495](https://github.com/getsentry/sentry-unreal/pull/1495))
   - [changelog](https://github.com/getsentry/sentry-android-gradle-plugin/blob/main/CHANGELOG.md#6160)
   - [diff](https://github.com/getsentry/sentry-android-gradle-plugin/compare/6.15.0...6.16.0)
+- Bump Native SDK from v0.15.4 to v0.16.0 ([#1499](https://github.com/getsentry/sentry-unreal/pull/1499))
+  - [changelog](https://github.com/getsentry/sentry-native/blob/master/CHANGELOG.md#0160)
+  - [diff](https://github.com/getsentry/sentry-native/compare/0.15.4...0.16.0)
 
 ## 1.17.0